refactor: Move shared e2e testkit to @chkit/clickhouse
Move ClickHouse-level test utilities (env, polling, naming) to
@chkit/clickhouse/e2e-testkit so all packages can import them.
CLI testkit re-exports and adds CLI-specific runner utilities.
