+++ packages/plugin-pull/src/pull.e2e.test.ts
@@ -4,90 +4,19 @@ import { join } from 'node:path'
 import { tmpdir } from 'node:os'
 
 import type { ResolvedChxConfig } from '@chkit/core'
+import { createClickHouseExecutor, type ClickHouseExecutor } from '@chkit/clickhouse'
+import {
+  getRequiredEnv,
+  quoteIdent,
+  createPrefix,
+  waitForTable,
+} from '@chkit/clickhouse/e2e-testkit'
 
 import { createPullPlugin } from './index.js'
 
-interface LiveEnv {
-  clickhouseUrl: string
-  clickhouseUser: string
-  clickhousePassword: string
-}
-
-function getRequiredEnv(): LiveEnv {
-  const clickhouseHost = process.env.CLICKHOUSE_HOST?.trim()
-  const clickhouseUrl =
-    process.env.CLICKHOUSE_URL?.trim() || (clickhouseHost ? `https://${clickhouseHost}` : '')
-  const clickhouseUser = process.env.CLICKHOUSE_USER?.trim() || 'default'
-  const clickhousePassword = process.env.CLICKHOUSE_PASSWORD?.trim() || ''
-
-  if (!clickhouseUrl) {
-    throw new Error('Missing CLICKHOUSE_URL or CLICKHOUSE_HOST')
-  }
-
-  if (!clickhousePassword) {
-    throw new Error('Missing CLICKHOUSE_PASSWORD')
-  }
-
-  return { clickhouseUrl, clickhouseUser, clickhousePassword }
-}
-
-function quoteIdent(value: string): string {
-  return '`' + value.replace(/`/g, '``') + '`'
-}
-
-async function runSql(url: string, username: string, password: string, sql: string): Promise<void> {
-  const auth = Buffer.from(`${username}:${password}`).toString('base64')
-  const response = await fetch(url, {
-    method: 'POST',
-    signal: AbortSignal.timeout(20_000),
-    headers: {
-      Authorization: `Basic ${auth}`,
-      'Content-Type': 'text/plain',
-    },
-    body: sql,
-  })
-
-  if (!response.ok) {
-    const body = await response.text()
-    throw new Error(`ClickHouse SQL failed (${response.status}): ${body}`)
-  }
-}
-
-async function querySql(url: string, username: string, password: string, sql: string): Promise<string> {
-  const auth = Buffer.from(`${username}:${password}`).toString('base64')
-  const response = await fetch(url, {
-    method: 'POST',
-    signal: AbortSignal.timeout(20_000),
-    headers: {
-      Authorization: `Basic ${auth}`,
-      'Content-Type': 'text/plain',
-    },
-    body: sql,
-  })
-
-  if (!response.ok) {
-    const body = await response.text()
-    throw new Error(`ClickHouse query failed (${response.status}): ${body}`)
-  }
-
-  return response.text()
-}
-
-async function dropDatabase(url: string, username: string, password: string, database: string): Promise<void> {
-  await runSql(url, username, password, `DROP DATABASE IF EXISTS ${quoteIdent(database)}`)
-}
-
-async function retry(attempts: number, delayMs: number, fn: () => Promise<void>): Promise<void> {
-  for (let i = 0; i < attempts; i++) {
-    try {
-      await fn()
-      return
-    } catch (err) {
-      if (i === attempts - 1) throw err
-      await new Promise((r) => setTimeout(r, delayMs))
-    }
-  }
-}
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
 
 function createResolvedConfig(clickhouse: NonNullable<ResolvedChxConfig['clickhouse']>): ResolvedChxConfig {
   return {
@@ -108,56 +37,40 @@ function createResolvedConfig(clickhouse: NonNullable<ResolvedChxConfig['clickho
   }
 }
 
-function createPrefix(label: string): string {
-  return `chkit_e2e_pull_${label}_${Date.now()}_${Math.floor(Math.random() * 100000)}_`
-}
-
 const runTag = process.env.GITHUB_RUN_ID?.trim() || `${Date.now()}_${Math.floor(Math.random() * 100000)}`
 const targetDatabase = `chkit_e2e_pull_${runTag}`
 const noiseDatabase = `chkit_e2e_pull_noise_${runTag}`
 
 describe('@chkit/plugin-pull live env e2e', () => {
   const liveEnv = getRequiredEnv()
+  let executor: ClickHouseExecutor
 
   beforeAll(async () => {
-    await retry(5, 1000, async () => {
-      await runSql(
-        liveEnv.clickhouseUrl,
-        liveEnv.clickhouseUser,
-        liveEnv.clickhousePassword,
-        `CREATE DATABASE IF NOT EXISTS ${quoteIdent(targetDatabase)}`
-      )
-      await runSql(
-        liveEnv.clickhouseUrl,
-        liveEnv.clickhouseUser,
-        liveEnv.clickhousePassword,
-        `CREATE DATABASE IF NOT EXISTS ${quoteIdent(noiseDatabase)}`
-      )
+    executor = createClickHouseExecutor({
+      url: liveEnv.clickhouseUrl,
+      username: liveEnv.clickhouseUser,
+      password: liveEnv.clickhousePassword,
+      database: 'default',
     })
+
+    await executor.execute(`CREATE DATABASE IF NOT EXISTS ${quoteIdent(targetDatabase)}`)
+    await executor.execute(`CREATE DATABASE IF NOT EXISTS ${quoteIdent(noiseDatabase)}`)
   })
 
   afterAll(async () => {
     try {
-      await dropDatabase(
-        liveEnv.clickhouseUrl,
-        liveEnv.clickhouseUser,
-        liveEnv.clickhousePassword,
-        targetDatabase
-      )
+      await executor.execute(`DROP DATABASE IF EXISTS ${quoteIdent(targetDatabase)}`)
     } catch {
       // Best-effort cleanup for shared e2e environment.
     }
 
     try {
-      await dropDatabase(
-        liveEnv.clickhouseUrl,
-        liveEnv.clickhouseUser,
-        liveEnv.clickhousePassword,
-        noiseDatabase
-      )
+      await executor.execute(`DROP DATABASE IF EXISTS ${quoteIdent(noiseDatabase)}`)
     } catch {
       // Best-effort cleanup for shared e2e environment.
     }
+
+    await executor.close()
   })
 
   test('validates required env variables are present', () => {
@@ -167,7 +80,7 @@ describe('@chkit/plugin-pull live env e2e', () => {
   test(
     'pulls table fixtures from dedicated database and excludes out-of-scope objects',
     async () => {
-      const prefix = createPrefix('objects')
+      const prefix = createPrefix('pull_objects')
       const eventsTable = `${prefix}events`
       const accountsTable = `${prefix}accounts`
       const eventsView = `${prefix}events_view`
@@ -188,66 +101,32 @@ describe('@chkit/plugin-pull live env e2e', () => {
       }
 
       try {
-        // ClickHouse Cloud DDL is eventually consistent; retry the first table creation.
-        await retry(3, 2000, () =>
-          runSql(
-            liveEnv.clickhouseUrl,
-            liveEnv.clickhouseUser,
-            liveEnv.clickhousePassword,
-            `CREATE TABLE ${quoteIdent(targetDatabase)}.${quoteIdent(eventsTable)} (id UInt64, source String, received_at DateTime64(3) DEFAULT now64(3)) ENGINE = MergeTree() PARTITION BY toYYYYMM(received_at) PRIMARY KEY (id) ORDER BY (id) SETTINGS index_granularity = 8192`
-          )
+        await executor.execute(
+          `CREATE TABLE ${quoteIdent(targetDatabase)}.${quoteIdent(eventsTable)} (id UInt64, source String, received_at DateTime64(3) DEFAULT now64(3)) ENGINE = MergeTree() PARTITION BY toYYYYMM(received_at) PRIMARY KEY (id) ORDER BY (id) SETTINGS index_granularity = 8192`
         )
-        await runSql(
-          liveEnv.clickhouseUrl,
-          liveEnv.clickhouseUser,
-          liveEnv.clickhousePassword,
+
+        // Wait for table to be visible before creating dependent objects
+        await waitForTable(executor, targetDatabase, eventsTable)
+
+        await executor.execute(
           `CREATE TABLE ${quoteIdent(targetDatabase)}.${quoteIdent(accountsTable)} (id UInt64, email Nullable(String), updated_at DateTime DEFAULT now()) ENGINE = MergeTree() PRIMARY KEY (id) ORDER BY (id)`
         )
-        await runSql(
-          liveEnv.clickhouseUrl,
-          liveEnv.clickhouseUser,
-          liveEnv.clickhousePassword,
+        await executor.execute(
           `CREATE VIEW ${quoteIdent(targetDatabase)}.${quoteIdent(eventsView)} AS SELECT id, source FROM ${quoteIdent(targetDatabase)}.${quoteIdent(eventsTable)}`
         )
-        await runSql(
-          liveEnv.clickhouseUrl,
-          liveEnv.clickhouseUser,
-          liveEnv.clickhousePassword,
+        await executor.execute(
           `CREATE TABLE ${quoteIdent(targetDatabase)}.${quoteIdent(eventsRollupTable)} (id UInt64, c UInt64) ENGINE = MergeTree() ORDER BY (id)`
         )
 
-        // ClickHouse Cloud DDL is eventually consistent; TO targets may lag visibility.
-        await retry(5, 2000, () =>
-          runSql(
-            liveEnv.clickhouseUrl,
-            liveEnv.clickhouseUser,
-            liveEnv.clickhousePassword,
-            `CREATE MATERIALIZED VIEW ${quoteIdent(targetDatabase)}.${quoteIdent(eventsMaterializedView)} TO ${quoteIdent(targetDatabase)}.${quoteIdent(eventsRollupTable)} AS SELECT id, count() AS c FROM ${quoteIdent(targetDatabase)}.${quoteIdent(eventsTable)} GROUP BY id`
-          )
-        )
-        await retry(3, 2000, () =>
-          runSql(
-            liveEnv.clickhouseUrl,
-            liveEnv.clickhouseUser,
-            liveEnv.clickhousePassword,
-            `CREATE TABLE ${quoteIdent(noiseDatabase)}.${quoteIdent(noiseTable)} (id UInt64) ENGINE = MergeTree() ORDER BY (id)`
-          )
-        )
+        // Wait for rollup table before creating MV that depends on it
+        await waitForTable(executor, targetDatabase, eventsRollupTable)
 
-        // ClickHouse Cloud DDL is eventually consistent — wait until all 5 objects
-        // (2 tables + 1 view + 1 rollup table + 1 materialized view) are visible.
-        await retry(15, 2000, async () => {
-          const result = await querySql(
-            liveEnv.clickhouseUrl,
-            liveEnv.clickhouseUser,
-            liveEnv.clickhousePassword,
-            `SELECT count() FROM system.tables WHERE database = '${targetDatabase}'`
-          )
-          const count = parseInt(result.trim(), 10)
-          if (count < 5) {
-            throw new Error(`Waiting for DDL visibility: expected 5 objects in ${targetDatabase}, found ${count}`)
-          }
-        })
+        await executor.execute(
+          `CREATE MATERIALIZED VIEW ${quoteIdent(targetDatabase)}.${quoteIdent(eventsMaterializedView)} TO ${quoteIdent(targetDatabase)}.${quoteIdent(eventsRollupTable)} AS SELECT id, count() AS c FROM ${quoteIdent(targetDatabase)}.${quoteIdent(eventsTable)} GROUP BY id`
+        )
+        await executor.execute(
+          `CREATE TABLE ${quoteIdent(noiseDatabase)}.${quoteIdent(noiseTable)} (id UInt64) ENGINE = MergeTree() ORDER BY (id)`
+        )
 
         const output: unknown[] = []
         const exitCode = await command.run({
